myDocker/requirements.txt
Flask
pillow
tensorflow
keras==2.2.4
cassandra-driver